move components to their own directories
I wanted to add a little more order to the component folder structure. This still feels somewhat inconsistent, but it at least is an improvement so styles, tests, and local constants, utils, hooks, could be colocated. In the future I think I should move locally used hooks and utils to these folders instead of the global hooks and utils dirs unless they are truly universal functions.